pyproject.toml
[project]
name = "repl-nix-workspace-fixed"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.69.0",
    "email-validator>=2.3.0",
    "flask>=3.1.2",
    "flask-dance[sqla]>=7.1.0",
    "flask-login>=0.6.3",
    "flask-sqlalchemy>=3.1.1",
    "gunicorn>=23.0.0",
    "psycopg2-binary>=2.9.11",
    "pyjwt>=2.10.1",
    "pytz>=2025.2",
    "requests>=2.32.5",
    "telethon>=1.41.2",
    "werkzeug>=3.1.3",
]
